[WEB] move Scenario to app module
