+++ app/build.gradle
@@ -32,7 +32,8 @@ def versions = [
         slf4j       : project.properties['slf4j.version'] ?: "2.0.3",
         gson        : project.properties['gson.version'] ?: "2.10",
         "commons-io": project.properties['commons-io.version'] ?: "2.11.0",
-        "opencsv"   : project.properties['opencsv.version'] ?: "5.6"
+        "opencsv"   : project.properties['opencsv.version'] ?: "5.6",
+        "math3"     : project.properties['math3.version'] ?: "3.6.1"
 ]
 
 dependencies {
@@ -51,6 +52,7 @@ dependencies {
     implementation "org.slf4j:slf4j-nop:${versions["slf4j"]}"
     implementation "com.google.code.gson:gson:${versions["gson"]}"
     implementation "com.opencsv:opencsv:${versions["opencsv"]}"
+    implementation "org.apache.commons:commons-math3:${versions["math3"]}"
 }
 
 application {
