+++ app/src/main/java/org/astraea/app/web/SkewedPartitionScenario.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.astraea.common.scenario;
+package org.astraea.app.web;
 
 import java.time.Duration;
 import java.util.ArrayList;
