+++ app/src/test/java/org/astraea/app/web/SkewedPartitionScenarioTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.astraea.common.scenario;
+package org.astraea.app.web;
 
 import java.util.List;
 import java.util.Set;
