+++ common/build.gradle
@@ -14,7 +14,6 @@ def versions = [
         mockito     : project.properties['mockito.version'] ?: "4.8.1",
         slf4j       : project.properties['slf4j.version'] ?: "2.0.3",
         "commons-io": project.properties['commons-io.version'] ?: "2.11.0",
-        "math3"     : project.properties['math3.version'] ?: "3.6.1",
         jackson     : project.properties['jackson.version'] ?: "2.13.4",
         "opencsv"   : project.properties['opencsv.version'] ?: "5.6"
 ]
@@ -31,7 +30,6 @@ dependencies {
     implementation "com.fasterxml.jackson.datatype:jackson-datatype-jdk8:${versions["jackson"]}"
     // we don't use slf4j actually, and it is used by kafka so we swallow the log.
     implementation "org.slf4j:slf4j-nop:${versions["slf4j"]}"
-    implementation "org.apache.commons:commons-math3:${versions["math3"]}"
     implementation "com.opencsv:opencsv:${versions["opencsv"]}"
 }
